Modify paste body storage
In order to facilitate atomic changes to pastes, paste files
are now named by a revision UUID, which is stored in
the database. When a paste is modified, the new body
is written to a file with a new revision UUID, and the
revision in the database is replaced atomically. The old
paste body is then queued to be deleted later by the
Manager.

Database schema updated to include additional
revision column.